pyproject.toml
[project]
name = "weather-insights-agent"
version = "0.1.0"
description = "Weather Insights & Forecast Advisor - Multi-agent system for emergency weather intelligence"
readme = "README.md"
requires-python = ">=3.10,<3.13"
dependencies = [
    "google-adk>=1.10.0",
    "a2a-sdk>=0.2.5",
    "google-cloud-bigquery",
    "google-cloud-logging",
    "python-dotenv",
    "requests",
    "uvicorn",
    "fastapi",
    "nest-asyncio>=1.6.0",
    "httpx",
    "google-generativeai",
    "folium",
]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["agents*", "shared_tools*"]
py-modules = []

[tool.setuptools.package-data]
weather_insights_agent = ["**/*.py"]
